requirements.txt
streamlit
PyPDF2
faiss-cpu
python-dotenv
google-generativeai
langchain
langchain_community
langchain_google_genai
